swagger

+++ src/controller/blogController.ts
@@ -10,10 +10,11 @@ export const createBlog = async (req: express.Request, res: express.Response) =>
         const user = await getUser((req as any).myAppToken);
         const {error} = joinBlogValidation(req.body);
         if(error){
-            
+            console.error(error);
+            res.status(400).json({ error: error.details[0].message });
         }
         // console.log(user)
-        if (user && user.isAdmin) {
+        if (user && user.isAdmin) { 
             // Create the blog
             Blog.create(req.body)
                 .then(blog => {
@@ -38,40 +39,37 @@ export const createBlog = async (req: express.Request, res: express.Response) =>
 export const fetchBlog = async(req: express.Request, res: express.Response)=>{
     try {
         const blogs = await Blog.find({});
-        res.status(200).json(blogs);
+        if(blogs.length === 0){
+            res.status(404).json({message:"THERE IS NO BLOG TO DISPLAY"})
+        }else{
+            res.status(200).json(blogs);
+        }
+        
     } catch (error) {
         console.log((error as Error).message);
-        res.status(500).json({message:(error as Error).message});
+        res.status(500).json({message:(error as Error).message});   
     }
 }
 
 //fetch blog by id
 export const fetchBlogById = async(req: express.Request, res: express.Response)=>{
     try {
-        const {id} = req.params;
-        const blog = await Blog.findById(id);
-        res.status(200).json(blog);
+        const checkUser = await getUser((req as any).myAppToken);
+        if(checkUser){
+            const {id} = req.params;
+            const blog = await Blog.findById(id);
+            res.status(200).json(blog);
+        } else{
+            res.status(401).json({ message: "YOU NEED TO LOGIN FIRST" });
+        }
+        
     } catch (error) {
         console.log((error as Error).message);
         res.status(500).json({message:(error as Error).message});
     }
 }
 
-//fetch blog by user id and blog id
-export const fetchBlogByUserIdAndBlogId = async(req: express.Request, res: express.Response)=>{
-    try {
-        const {userId} = req.params;
-        const {blogId} = req.params;
-        const blog = await Blog.findOne({userId: userId, _id: blogId},req.body);
-        if(!blog){
-            return res.status(404).json({message:'Blog not found for the specified userID and blogID'});
-        }
-        res.status(200).json(blog);
-    } catch (error) {
-        console.log((error as Error).message);
-        res.status(500).json({message:(error as Error).message});
-    }
-}
+
 
 //patch blog by id
 export const patchBlogById = async(req: express.Request, res: express.Response)=>{
@@ -99,12 +97,18 @@ export const patchBlogById = async(req: express.Request, res: express.Response)=
 //Delete blog by id
 export const deleteBlog = async(req: express.Request, res: express.Response)=>{
     try {
-        const {id} = req.params;
-        const blog = await Blog.deleteOne({ _id: id });
-        if(blog.deletedCount === 0){
-            return res.status(404).json({message:`cannot find any category with ID ${id}`});
+        const checkUser = await getUser((req as any).myAppToken);
+        if(checkUser && checkUser.isAdmin){
+            const {id} = req.params;
+            const blog = await Blog.deleteOne({ _id: id });
+            if(blog.deletedCount === 0){
+                return res.status(404).json({message:`cannot find any category with ID ${id}`});
+            }
+            res.status(200).json({blog,message:"Blog deleted successfully "});
+        } else{
+            res.status(401).json({ message: "YOU ARE NOT AUTHORIZED TO DELETE A BLOG" });
         }
-        res.status(200).json(blog);
+       
     } catch (error) {
         console.log((error as Error).message);
         res.status(500).json({message:(error as Error).message});
